+++ src/main/java/com/lama/sc/utils/Utils.java
@@ -27,7 +27,7 @@ public final class Utils {
 		return SEED.nextInt((max - min) + 1) + min;
 	}
 	
-	public static int log2(int value){
+	public static long log2(long value){
 		return (int) (Math.log10(value) / Math.log10(2));
 	}
